+++ src/main/java/com/taskManager/taskManager/entities/Project.java
@@ -18,7 +18,7 @@ public class Project {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "project_id", nullable = false)
-	private Long projectId;
+	private int projectId;
 
 	@Column(name = "project_name", nullable = false, length = 50)
 	@NotBlank(message = "Project name is required.")
